-Incialice la variable i para que ande bien la lineas 29 unlist -Añadi la variable download, para solamente cargar datos

+++ src/load_data.R
@@ -1,5 +1,7 @@
-load_data <- function()
+load_data <- function(download)
 {
+	if(missing(download))download = FALSE
+	
 	#Esta funcion parsea el csv producido por ivco19 Libraries
 	#https://github.com/ivco19/libs
 	#descarga la base de datos en formato csv con los datos de Argentina
@@ -8,22 +10,22 @@ load_data <- function()
 	#y un archivo minimal_data.dat con los totales a nivel nacional del número
 	#de casos confirmados, recuperados, activos y muertos por dia. En la salida 
 	#la función devuelve un data.frame con las series temporales de estos totales.
+ 	if(download){
+           #borrado de posible version anterior
+	   system("rm pub?output=csv")
  
-        #borrado de posible version anterior
-	system("rm pub?output=csv")
- 
-	#donwload
-	system("wget https://docs.google.com/spreadsheets/d/e/2PACX-1vTfinng5SDBH9RSJMHJk28dUlW3VVSuvqaBSGzU-fYRTVLCzOkw1MnY17L2tWsSOppHB96fr21Ykbyv/pub?output=csv")
-        #renombrado a data.csv
-	system("rm data.csv")
-	system("mv pub?output=csv data.csv")
-
+	   #donwload
+	   system("wget https://docs.google.com/spreadsheets/d/e/2PACX-1vTfinng5SDBH9RSJMHJk28dUlW3VVSuvqaBSGzU-fYRTVLCzOkw1MnY17L2tWsSOppHB96fr21Ykbyv/pub?output=csv")
+           #renombrado a data.csv
+	   system("rm data.csv")
+	   system("mv pub?output=csv data.csv")
+	}
 	#numero de filas (24 (provincias)* 4 (muertos, confirmados, recuperados y activos)
 	Nf=96
 
 
        	dm = read.csv2("data.csv",stringsAsFactors=FALSE)
-
+	i=1
         line=unlist(strsplit(dm[[1]][i],","))
 	#numero de días desde el primer caso confirmado
 	Nd=length(line)
